+++ lib/core.dart
@@ -2,13 +2,11 @@ library;
 
 export 'src/core/configs.dart';
 export 'src/core/extensions.dart';
-export 'src/models/checker.dart';
-export 'src/models/updating_info.dart';
-export 'src/repositories/base.dart';
-export 'src/repositories/local.dart';
-export 'src/repositories/remote.dart';
+export 'src/core/checker.dart';
+export 'src/core/updating_info.dart';
+export 'src/core/repository.dart';
 export 'src/sources/base.dart';
 export 'src/sources/local.dart';
 export 'src/sources/remote.dart';
-export 'src/utils/encryptor.dart';
-export 'src/utils/errors.dart';
+export 'src/core/encryptor.dart';
+export 'src/core/errors.dart';