+++ lib/data_management.dart
@@ -1,6 +1,4 @@
 library;
 
-export 'package:flutter_entity/flutter_entity.dart';
-
 export 'core.dart';
 export 'widgets.dart';